backend/requirements.txt
flask
flask-socketio
flask-cors
python-socketio
flask-sqlalchemy
python-dotenv
